Add an executable node which will use PCL
Point Cloud Library (PCL) which is integrated into ros, is going to be
used for point cloud manipulation. This ROS node currently just echoes
the point cloud recieved by the depth camera, but will in the future
edit the point cloud to remove out-of-bounds objects, de-noise it, and
try to fit a ball model to the point cloud.
 Changes to be committed:
	modified:   ../CMakeLists.txt
	modified:   ../package.xml
	new file:   point_cloud_detect.cpp
	new file:   point_cloud_detect.h